Add files via upload
This dataset provides a comprehensive depiction of the retail landscape. It encompasses sales data for an extensive array of products tracked across distinct stores situated in diverse cities across 12 variables in this dataset. For this project, please use the aforementioned link. You may use the ‘train.csv’ and ‘test.csv’ for training and testing the model respectively. Delving into the complex features associated with each product and store, the dataset captures the dynamics of sales for each product and defines store characteristics. This includes details on store size, location specifics, and other pertinent attributes, offering a detailed exploration of the retail environment. This dataset offers a realistic simulation of real-world data analysis challenges.